refactor: move the truncation gadgets into composer::truncate
Pure code motion: `component_truncate`, `bind_truncation_split` and
`assert_canonical_truncation` move out of `composer.rs` into a
`truncate` submodule, together with their soundness tests. The two
helpers become `pub(super)` so the logic gadget and its tests can still
reach them.
